requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
Pygments==2.18.0
accelerate==1.1.1
aiofiles==23.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
audioread==3.0.1
av==13.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
editdistance==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
funasr==1.1.14
future==1.0.0
gradio==5.6.0
gradio_client==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
huggingface==0.0.1
hydra-core==1.3.2
idna==3.10
imageio-ffmpeg==0.5.1
imageio==2.36.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
modelscope==1.20.1
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
omegaconf==2.3.0
openai==1.58.1
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.12
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytorch-wpe==0.0.1
pytube==15.0.0
pytz==2024.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.8.0
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.3
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomlkit==0.12.0
torch-complex==0.4.4
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
uvicorn==0.32.1
websockets==12.0
wheel==0.44.0
